Update Min steps to gold.cpp

+++ CPP/Min steps to gold.cpp
@@ -1,6 +1,6 @@
 int Solution::coverPoints(vector<int> &A, vector<int> &B) {
-    int x=A[0], y=B[0], sum=0;
-    for(int i=1;i<A.size();i++)
+    long long x=A[0], y=B[0], sum=0;
+    for(long long i=1;i<A.size();i++)
     {
         if(abs(B[i]-y)>abs(A[i]-x))
         {
